fix a soundness bug in the project! macro
I've made one of the classic mistakes of unsafe code: I created a
reference unsafely, and I didn't bind its lifetime to anything. That
made it possible for the caller to do something wild like this

    let y = {
        let x = Cell::new(5);
        let x = &x;
        project!(x)
    };

Woops! Now x has been destroyed, and using y triggers UB and fails Miri.
To work around this, we need to let the compiler know that the
references we create are bound to something, which we can do by passing
them through a helper function.